+++ tests/e2e/pipelex/pipes/pipe_controller/pipe_condition/test_pipe_condition_2.py
@@ -9,7 +9,7 @@ from pipelex.core.stuff_factory import StuffFactory
 from pipelex.core.working_memory_factory import WorkingMemoryFactory
 from pipelex.hub import get_pipe_router, get_report_delegate
 from pipelex.pipeline.job_metadata import JobMetadata
-from tests.pipelines.pipe_controllers.pipe_condition.pipe_condition import CategoryInput
+from tests.test_pipelines.pipe_controllers.pipe_condition.pipe_condition import CategoryInput
 
 
 @pytest.mark.llm
