+++ tests/integration/pipelex/pipes/pipe_controller/pipe_condition/test_pipe_condition_simple.py
@@ -17,7 +17,7 @@ from pipelex.exceptions import DryRunError
 from pipelex.hub import get_pipe_router
 from pipelex.pipe_controllers.pipe_condition import PipeCondition
 from pipelex.pipeline.job_metadata import JobMetadata
-from tests.pipelines.pipe_controllers.pipe_condition.pipe_condition import CategoryInput
+from tests.test_pipelines.pipe_controllers.pipe_condition.pipe_condition import CategoryInput
 
 
 @pytest.mark.dry_runnable
