+++ tests/test_pipelines/misc_tests/test_image_inputs.toml
