+++ tests/test_pipelines/pipe_controllers/pipe_sequence/capitalize_text.toml
