+++ tests/test_pipelines/tricky_questions.py
